# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "An API Boilerplate",
    "title": "Renco Boilerplate",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/license/MIT"
    },
    "version": "0.1"
  },
  "basePath": "/api/v1",
  "paths": {
    "/auth/access-token": {
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "Get access token",
        "parameters": [
          {
            "description": "email and password",
            "name": "credential",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/domain.AuthControllerPayloadGetAccessToken"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "validation error"
          },
          "404": {
            "description": "user not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "Create user",
        "parameters": [
          {
            "description": "email and password",
            "name": "credential",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/domain.AuthControllerPayloadSignUp"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "validation error | user already exist"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "domain.AuthControllerPayloadGetAccessToken": {
      "type": "object",
      "required": ["email", "password"],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string",
          "minLength": 8
        }
      }
    },
    "domain.AuthControllerPayloadSignUp": {
      "type": "object",
      "required": ["email", "is_admin", "password"],
      "properties": {
        "email": {
          "type": "string"
        },
        "is_admin": {
          "type": "boolean"
        },
        "password": {
          "type": "string",
          "minLength": 8
        }
      }
    }
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "description": "Firebase auth access token, get it from POST /auth/access-token",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}